backend/pyproject.toml
[tool.poetry]
name = "senior-citizen-backend"
version = "1.0.0"
description = "Flask API for the senior-citizen app"
authors = ["RNA <your.email@example.com>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "backend" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
flask = "^3.1.1"
gunicorn = ">=23.0.0,<24.0.0"
sqlalchemy = ">=1.4"
flask_jwt_extended = "^4.6.0"
bcrypt = "^4.1.2"
flasgger = "^0.9.7.1"
flask-smorest = "^0.44.0"
marshmallow = "^3.20.2"
apispec = "^6.3.0"
apispec-webframeworks = "^0.5.2"
flask-security-too = "^5.6.2"
flask-sqlalchemy = "^3.1.1"
apscheduler = "^3.11.0"
twilio = "^9.6.4"
authlib = "^1.6.0"
python-dotenv = "^1.1.1"
flask-cors = "^6.0.1"
requests = "^2.32.4"
redis = "^5.0.1"
celery = "^5.5.3"
flask-socketio = "^5.5.1"
flask-mail = "^0.10.0"
flask-principal = { git = "https://github.com/mattupstate/flask-principal" }
bleach = "^6.1.0"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
black = "*"
ruff = "*"
mypy = "^1.13.0"
pytest-mock = "^3.14.1"
types-bleach = "^6.2.0.20250514"
types-pytz = "*"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-flask = "^1.3.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
python_version = "3.11"
strict = true
# Remove this if not using Pydantic:
# plugins = ["pydantic.mypy"]
